+++ scripts/data_prep/extract_images.py
@@ -53,15 +53,19 @@ def main(data_dir):
             print('Extracting image for sequence {}...'.format(s.seq_id))
             img, _ = extract_image(s)
 
-            # Write image to disk.
-            try:
-                cv2.imwrite(output_fn, img)
-                extract_count += 1
-            except KeyboardInterrupt as e:
-                # Safely handle premature termination.  Remove unfinished file.
-                if os.exists(output_fn):
-                    os.remove(output_fn)
-                raise e
+            if img is None:
+                print("    No frame with a face found")
+            else:
+                # Write image to disk.
+                try:
+                    cv2.imwrite(output_fn, img)
+                    extract_count += 1
+                except KeyboardInterrupt as e:
+                    # Safely handle premature termination.
+                    # Remove unfinished file.
+                    if os.exists(output_fn):
+                        os.remove(output_fn)
+                    raise e
 
     except KeyboardInterrupt:
         print('Program terminated prematurely')
